loplugin:redundantcast: vcl
(after a to-be-committed improved loplugin:cstylecast would have rewritten the
C-style casts into static_casts)

+++ vcl/source/gdi/animate.cxx
@@ -48,7 +48,7 @@ BitmapChecksum AnimationBitmap::GetChecksum() const
     UInt32ToSVBT32( aSizePix.Height(), aBT32 );
     nCrc = vcl_get_checksum( nCrc, aBT32, 4 );
 
-    UInt32ToSVBT32( (long) nWait, aBT32 );
+    UInt32ToSVBT32( nWait, aBT32 );
     nCrc = vcl_get_checksum( nCrc, aBT32, 4 );
 
     UInt32ToSVBT32( (long) eDisposal, aBT32 );
